refactor tx limiting

+++ src/transaction.rs
@@ -5,12 +5,12 @@ pub struct Transaction {
     sender: String,
     receiver: String,
     amount: i64, // TODO: turn into floating point value
-    pub birth: i32,
+    pub birth: usize,
     // TODO: add tip
 }
 
 impl Transaction {
-    pub fn new(sender: String, receiver: String, amount: i64, birth: i32) -> Self {
+    pub fn new(sender: String, receiver: String, amount: i64, birth: usize) -> Self {
         Self {
             sender,
             receiver,
